+++ frontend/src/locales/zh-CN.json
@@ -601,11 +601,18 @@
       "global-masking-rule": "全局脱敏规则",
       "data-feature": "数据特征",
       "add-rule": "添加规则",
+      "grant-access": "授权访问",
+      "never-expires": "无过期时间",
       "masking-level": {
         "self": "脱敏程度",
         "none": "明文",
         "partial": "半脱敏",
         "full": "全脱敏"
+      },
+      "action": {
+        "self": "操作类型",
+        "query": "查询",
+        "export": "导出"
       }
     },
     "access-control": {
